Reorganized Dialogs in Kanban
