+++ src/components/KanbanBoard/ListMoreButton.js
@@ -18,8 +18,8 @@ import MoreIcon from '@material-ui/icons/MoreVert';
 import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
 import FileCopyOutlinedIcon from '@material-ui/icons/FileCopyOutlined';
 import DeleteOutlineIcon from '@material-ui/icons/DeleteOutline';
-import EditTaskDialog from './EditTaskDialog';
-import DeleteTaskDialog from './DeleteTaskDialog'
+import EditTaskDialog from './TaskDialogs/EditTaskDialog';
+import DeleteTaskDialog from './TaskDialogs/DeleteTaskDialog'
 
 function ListMoreButton(props) {
   const moreRef = useRef(null);
