+++ src/components/KanbanBoard/TaskDialogs/DeleteTaskDialog.js
